requirements.txt
mcp>=1.13.1
yfinance>=0.2.65
pydantic>=2.11.7
requests>=2.31.0
python-dotenv>=1.0.0
